[DFS] Add API to write byte array data without temporary files, fix section size in card manager

+++ dfs/controls/sources/dfs.cpp
@@ -51,19 +51,31 @@ void Dfs::initDFS(const QByteArray &userId)
     qDebug() << "[init finished]";
 }
 
-void Dfs::saveToDFS(const QString &path, const dfsStruct::Type &type, const dfsStruct::SubType &subType,
-                    const dfsStruct::Status &status)
+void Dfs::saveToDFS(const QString &path, const QByteArray &data, const dfsStruct::Type &type,
+                    const dfsStruct::SubType &subType, const dfsStruct::Status &status)
 {
-    QFile file(path);
     QByteArray userId = accountControler->getMainActor()->getId().toActorId();
     QByteArray dfsPath = buildDfsPath(userId, type);
     if (!appendToCard(dfsPath, userId, type, subType))
         return;
-    if (!file.copy(dfsPath))
+
+    if (path.isEmpty())
     {
-        QFile::remove(dfsPath);
-        file.copy(dfsPath);
+        QFile file(dfsPath);
+        file.open(QFile::WriteOnly);
+        file.write(data);
+        file.close();
     }
+    else
+    {
+        QFile file(path);
+        if (!file.copy(dfsPath))
+        {
+            QFile::remove(dfsPath);
+            file.copy(dfsPath);
+        }
+    }
+
     sender->sendFile(dfsPath, type, SocketPair());
 #ifdef ETALONIUM_CLIENT
     emit usersChanges(dfsPath, type, userId); // TODO
@@ -310,23 +322,24 @@ void Dfs::initDFSNetManager(ResolveManager *resolveManager)
     ThreadPool::addThread(dfsNetManager);
 }
 
-void Dfs::savedNewData(const QString &path, const dfsStruct::Type &type, const dfsStruct::SubType &subType,
-                       const dfsStruct::Status &status)
+void Dfs::savedNewData(const QString &path, const QByteArray &data, const dfsStruct::Type &type,
+                       const dfsStruct::SubType &subType, const dfsStruct::Status &status)
 {
 #ifdef ETALONIUM_CLIENT
     if (type == dfsStruct::Type::images && subType != dfsStruct::SubType::mini)
     {
-        QImage im(path);
-        if (im.save("temp", "jpeg", 80))
+        QImage image(path);
+
+        if (image.save("temp_image", "jpeg", 80))
         {
-            saveToDFS("temp", type, subType, status);
-            //            QFile filed("temp");
-            //            filed.remove();
+            saveToDFS("temp_image", "", type, subType, status);
+            QFile::remove("temp_image");
             return;
         }
     }
 #endif
-    saveToDFS(path, type, subType, status);
+
+    saveToDFS(path, data, type, subType, status);
 }
 
 void Dfs::process()
